Added Functionality: 8. Airdrop the same NFT to any users with different amounts
added memos for Nft transfers
when airdropping/transferring a count will display showing where you are in the process instead of the storage id and offchain fee
issue with the input.txt fixed. you can now change the file before hitting yes and not have to restart the application after using a functionality and then moving onto another.